Refactor TTS client
